+++ NEWS.rst
@@ -6,7 +6,10 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Update kodo to version 25.
+* Major: Update fifi to version 19.
+* Major: Rename ``encode``, ``decode``, and ``recode`` to ``write_payload``,
+  ``read_payload``, and ``write_payload``, respectively.
 
 6.0.2
 -----